Split HorseshoeTool, SegArcTool, BarChartTool
