+++ srcjs/horseshoetool.js
@@ -1,0 +1,308 @@
+import { svg } from 'lit-element';
+import { classMap } from 'lit-html/directives/class-map.js';
+import { styleMap } from 'lit-html/directives/style-map.js';
+
+import Merge from './merge';
+import BaseTool from './basetool';
+
+/** ****************************************************************************
+  * HorseshoeTool class
+  *
+  * Summary.
+  *
+  */
+
+export default class HorseshoeTool extends BaseTool {
+  // Donut starts at -220 degrees and is 260 degrees in size.
+  // zero degrees is at 3 o'clock.
+
+  constructor(argToolset, argConfig, argPos) {
+    const DEFAULT_HORSESHOE_CONFIG = {
+      position: {
+        cx: 50,
+        cy: 50,
+        radius: 45,
+      },
+      card_filter: 'card--filter-none',
+      horseshoe_scale: {
+        min: 0,
+        max: 100,
+        width: 3,
+        color: 'var(--primary-background-color)',
+      },
+      horseshoe_state: {
+        width: 6,
+        color: 'var(--primary-color)',
+      },
+      show: {
+        horseshoe: true,
+        scale_tickmarks: false,
+        horseshoe_style: 'fixed',
+      },
+    };
+
+    super(argToolset, Merge.mergeDeep(DEFAULT_HORSESHOE_CONFIG, argConfig), argPos);
+
+    // Next consts are now variable. Should be calculated!!!!!!
+    this.HORSESHOE_RADIUS_SIZE = 0.45 * SVG_VIEW_BOX;
+    this.TICKMARKS_RADIUS_SIZE = 0.43 * SVG_VIEW_BOX;
+    this.HORSESHOE_PATH_LENGTH = 2 * 260 / 360 * Math.PI * this.HORSESHOE_RADIUS_SIZE;
+
+    // this.config = {...DEFAULT_HORSESHOE_CONFIG};
+    // this.config = {...this.config, ...argConfig};
+
+    // if (argConfig.styles) this.config.styles = {...argConfig.styles};
+    // this.config.styles = {...DEFAULT_HORSESHOE_CONFIG.styles, ...this.config.styles};
+
+    // //if (argConfig.show) this.config.show = Object.assign(...argConfig.show);
+    // this.config.show = {...DEFAULT_HORSESHOE_CONFIG.show, ...this.config.show};
+
+    // //if (argConfig.horseshoe_scale) this.config.horseshoe_scale = Object.assign(...argConfig.horseshoe_scale);
+    // this.config.horseshoe_scale = {...DEFAULT_HORSESHOE_CONFIG.horseshoe_scale, ...this.config.horseshoe_scale};
+
+    // // if (argConfig.horseshoe_state) this.config.horseshoe_state = Object.assign(...argConfig.horseshoe_state);
+    // this.config.horseshoe_state = {...DEFAULT_HORSESHOE_CONFIG.horseshoe_state, ...this.config.horseshoe_state};
+
+    this.config.entity_index = this.config.entity_index ? this.config.entity_index : 0;
+
+    this.svg.radius = Utils.calculateSvgDimension(this.config.position.radius);
+    this.svg.radius_ticks = Utils.calculateSvgDimension(0.95 * this.config.position.radius);
+
+    this.svg.horseshoe_scale = {};
+    this.svg.horseshoe_scale.width = Utils.calculateSvgDimension(this.config.horseshoe_scale.width);
+    this.svg.horseshoe_state = {};
+    this.svg.horseshoe_state.width = Utils.calculateSvgDimension(this.config.horseshoe_state.width);
+    this.svg.horseshoe_scale.dasharray = 2 * 26 / 36 * Math.PI * this.svg.radius;
+
+    // The horseshoe is rotated around its svg base point. This is NOT the center of the circle!
+    // Adjust x and y positions within the svg viewport to re-center the circle after rotating
+    this.svg.rotate = {};
+    this.svg.rotate.degrees = -220;
+    this.svg.rotate.cx = this.svg.cx;
+    this.svg.rotate.cy = this.svg.cy;
+
+    // Get colorstops and make a key/value store...
+    this.colorStops = {};
+    if (this.config.color_stops) {
+      Object.keys(this.config.color_stops).forEach((key) => {
+        this.colorStops[key] = this.config.color_stops[key];
+      });
+    }
+
+    this.sortedStops = Object.keys(this.colorStops).map((n) => Number(n)).sort((a, b) => a - b);
+
+    // Create a colorStopsMinMax list for autominmax color determination
+    this.colorStopsMinMax = {};
+    this.colorStopsMinMax[this.config.horseshoe_scale.min] = this.colorStops[this.sortedStops[0]];
+    this.colorStopsMinMax[this.config.horseshoe_scale.max] = this.colorStops[this.sortedStops[(this.sortedStops.length) - 1]];
+
+    // Now set the color0 and color1 for the gradient used in the horseshoe to the colors
+    // Use default for now!!
+    this.color0 = this.colorStops[this.sortedStops[0]];
+    this.color1 = this.colorStops[this.sortedStops[(this.sortedStops.length) - 1]];
+
+    this.angleCoords = {
+      x1: '0%', y1: '0%', x2: '100%', y2: '0%',
+    };
+    // this.angleCoords = angleCoords;
+    this.color1_offset = '0%';
+
+    //= ===================
+    // End setConfig part.
+
+    if (this.dev.debug) console.log('HorseshoeTool constructor coords, dimensions', this.coords, this.dimensions, this.svg, this.config);
+  }
+
+  /** *****************************************************************************
+  * HorseshoeTool::value()
+  *
+  * Summary.
+  * Sets the value of the horseshoe. Value updated via set hass.
+  * Calculate horseshoe settings & colors depening on config and new value.
+  *
+  */
+
+  set value(state) {
+    if (this._stateValue === state) return false;
+
+    this._stateValuePrev = this._stateValue || state;
+    this._stateValue = state;
+    this._stateValueIsDirty = true;
+
+    // Calculate the size of the arc to fill the dasharray with this
+    // value. It will fill the horseshoe relative to the state and min/max
+    // values given in the configuration.
+
+    const min = this.config.horseshoe_scale.min || 0;
+    const max = this.config.horseshoe_scale.max || 100;
+    const val = Math.min(this._card._calculateValueBetween(min, max, state), 1);
+    const score = val * this.HORSESHOE_PATH_LENGTH;
+    const total = 10 * this.HORSESHOE_RADIUS_SIZE;
+    this.dashArray = `${score} ${total}`;
+
+    // We must draw the horseshoe. Depending on the stroke settings, we draw a fixed color, gradient, autominmax or colorstop
+    // #TODO: only if state or attribute has changed.
+
+    const strokeStyle = this.config.show.horseshoe_style;
+
+    if (strokeStyle === 'fixed') {
+      this.stroke_color = this.config.horseshoe_state.color;
+      this.color0 = this.config.horseshoe_state.color;
+      this.color1 = this.config.horseshoe_state.color;
+      this.color1_offset = '0%';
+      //  We could set the circle attributes, but we do it with a variable as we are using a gradient
+      //  to display the horseshoe circle .. <horseshoe circle>.setAttribute('stroke', stroke);
+    } else if (strokeStyle === 'autominmax') {
+      // Use color0 and color1 for autoranging the color of the horseshoe
+      const stroke = this._card._calculateColor(state, this.colorStopsMinMax, true);
+
+      // We now use a gradient for the horseshoe, using two colors
+      // Set these colors to the colorstop color...
+      this.color0 = stroke;
+      this.color1 = stroke;
+      this.color1_offset = '0%';
+    } else if (strokeStyle === 'colorstop' || strokeStyle === 'colorstopgradient') {
+      const stroke = this._card._calculateColor(state, this.colorStops, strokeStyle === 'colorstopgradient');
+
+      // We now use a gradient for the horseshoe, using two colors
+      // Set these colors to the colorstop color...
+      this.color0 = stroke;
+      this.color1 = stroke;
+      this.color1_offset = '0%';
+    } else if (strokeStyle === 'lineargradient') {
+      // This has taken a lot of time to get a satisfying result, and it appeared much simpler than anticipated.
+      // I don't understand it, but for a circle, a gradient from left/right with adjusted stop is enough ?!?!?!
+      // No calculations to adjust the angle of the gradient, or rotating the gradient itself.
+      // Weird, but it works. Not a 100% match, but it is good enough for now...
+
+      // According to stackoverflow, these calculations / adjustments would be needed, but it isn't ;-)
+      // Added from https://stackoverflow.com/questions/9025678/how-to-get-a-rotated-linear-gradient-svg-for-use-as-a-background-image
+      const angleCoords = {
+        x1: '0%', y1: '0%', x2: '100%', y2: '0%',
+      };
+      this.color1_offset = `${Math.round((1 - val) * 100)}%`;
+
+      this.angleCoords = angleCoords;
+    }
+    if (this.dev.debug) console.log('HorseshoeTool set value', this.cardId, state);
+
+    return true;
+  }
+
+  /** *****************************************************************************
+  * HorseshoeTool::_renderTickMarks()
+  *
+  * Summary.
+  * Renders the tick marks on the scale.
+  *
+  */
+
+  _renderTickMarks() {
+    const { config } = this;
+    // if (!config) return;
+    // if (!config.show) return;
+    if (!config.show.scale_tickmarks) return;
+
+    const stroke = config.horseshoe_scale.color ? config.horseshoe_scale.color : 'var(--primary-background-color)';
+    const tickSize = config.horseshoe_scale.ticksize ? config.horseshoe_scale.ticksize
+      : (config.horseshoe_scale.max - config.horseshoe_scale.min) / 10;
+
+    // fullScale is 260 degrees. Hard coded for now...
+    const fullScale = 260;
+    const remainder = config.horseshoe_scale.min % tickSize;
+    const startTickValue = config.horseshoe_scale.min + (remainder === 0 ? 0 : (tickSize - remainder));
+    const startAngle = ((startTickValue - config.horseshoe_scale.min)
+                        / (config.horseshoe_scale.max - config.horseshoe_scale.min)) * fullScale;
+    const tickSteps = ((config.horseshoe_scale.max - startTickValue) / tickSize);
+
+    // new
+    let steps = Math.floor(tickSteps);
+    const angleStepSize = (fullScale - startAngle) / tickSteps;
+
+    // If steps exactly match the max. value/range, add extra step for that max value.
+    if ((Math.floor(((steps) * tickSize) + startTickValue)) <= (config.horseshoe_scale.max)) { steps += 1; }
+
+    const radius = this.svg.horseshoe_scale.width ? this.svg.horseshoe_scale.width / 2 : 6 / 2;
+    let angle;
+    const scaleItems = [];
+
+    // NTS:
+    // Value of -230 is weird. Should be -220. Can't find why...
+    let i;
+    for (i = 0; i < steps; i++) {
+      angle = startAngle + ((-230 + (360 - i * angleStepSize)) * Math.PI / 180);
+      scaleItems[i] = svg`
+        <circle cx="${this.svg.cx - Math.sin(angle) * this.svg.radius_ticks}"
+                cy="${this.svg.cy - Math.cos(angle) * this.svg.radius_ticks}" r="${radius}"
+                fill="${stroke}">
+      `;
+    }
+    return svg`${scaleItems}`;
+  }
+
+  /** *****************************************************************************
+  * HorseshoeTool::_renderHorseShoe()
+  *
+  * Summary.
+  * Renders the horseshoe group.
+  *
+  * Description.
+  * The horseshoes are rendered in a viewbox of 200x200 (SVG_VIEW_BOX).
+  * Both are centered with a radius of 45%, ie 200*0.45 = 90.
+  *
+  * The foreground horseshoe is always rendered as a gradient with two colors.
+  *
+  * The horseshoes are rotated 220 degrees and are 2 * 26/36 * Math.PI * r in size
+  * There you get your value of 408.4070449,180 ;-)
+  */
+
+  _renderHorseShoe() {
+    if (!this.config.show.horseshoe) return;
+
+    return svg`
+      <g id="horseshoe__group-inner" class="horseshoe__group-inner">
+        <circle id="horseshoe__scale" class="horseshoe__scale" cx="${this.svg.cx}" cy="${this.svg.cy}" r="${this.svg.radius}"
+          fill="${this.fill || 'rgba(0, 0, 0, 0)'}"
+          stroke="${this.config.horseshoe_scale.color || '#000000'}"
+          stroke-dasharray="${this.svg.horseshoe_scale.dasharray}"
+          stroke-width="${this.svg.horseshoe_scale.width}"
+          stroke-linecap="square"
+          transform="rotate(-220 ${this.svg.rotate.cx} ${this.svg.rotate.cy})"/>
+
+        <circle id="horseshoe__state__value" class="horseshoe__state__value" cx="${this.svg.cx}" cy="${this.svg.cy}" r="${this.svg.radius}"
+          fill="${this.config.fill || 'rgba(0, 0, 0, 0)'}"
+          stroke="url('#horseshoe__gradient-${this.cardId}')"
+          stroke-dasharray="${this.dashArray}"
+          stroke-width="${this.svg.horseshoe_state.width}"
+          stroke-linecap="square"
+          transform="rotate(-220 ${this.svg.rotate.cx} ${this.svg.rotate.cy})"/>
+
+        ${this._renderTickMarks()}
+      </g>
+    `;
+  }
+
+  /** *****************************************************************************
+  * HorseshoeTool::render()
+  *
+  * Summary.
+  * The render() function for this object.
+  *
+  */
+  render() {
+    return svg`
+      <g "" id="horseshoe-${this.toolId}" class="horseshoe__group-outer"
+        @click=${(e) => this.handleTapEvent(e, this.config)}>
+        ${this._renderHorseShoe()}
+      </g>
+
+      <svg style="width:0;height:0;position:absolute;" aria-hidden="true" focusable="false">
+        <linearGradient gradientTransform="rotate(0)" id="horseshoe__gradient-${this.cardId}" x1="${this.angleCoords.x1}", y1="${this.angleCoords.y1}", x2="${this.angleCoords.x2}" y2="${this.angleCoords.y2}">
+          <stop offset="${this.color1_offset}" stop-color="${this.color1}" />
+          <stop offset="100%" stop-color="${this.color0}" />
+        </linearGradient>
+      </svg>
+
+    `;
+  }
+} // END of class
