+++ srcjs/sparklinebarcharttool.js
@@ -1,0 +1,234 @@
+import { svg } from 'lit-element';
+import { classMap } from 'lit-html/directives/class-map.js';
+import { styleMap } from 'lit-html/directives/style-map.js';
+
+import Merge from './merge';
+import BaseTool from './basetool';
+import Utils from './utils';
+
+/** ****************************************************************************
+  * SparklineBarChartTool class
+  *
+  * Summary.
+  *
+  */
+export default class SparklineBarChartTool extends BaseTool {
+  constructor(argToolset, argConfig, argPos) {
+    const DEFAULT_BARCHART_CONFIG = {
+      position: {
+        cx: 50,
+        cy: 50,
+        height: 25,
+        width: 25,
+        margin: 0.5,
+        orientation: 'vertical',
+      },
+      hours: 24,
+      barhours: 1,
+      color: 'var(--primary-color)',
+      classes: {
+        tool: {
+          'sak-barchart': true,
+          hover: true,
+        },
+        bar: {
+        },
+        line: {
+          'sak-barchart__line': true,
+          hover: true,
+        },
+      },
+      styles: {
+        tool: {
+        },
+        line: {
+        },
+        bar: {
+        },
+      },
+      colorstops: [],
+      show: { style: 'fixedcolor' },
+    };
+
+    super(argToolset, Merge.mergeDeep(DEFAULT_BARCHART_CONFIG, argConfig), argPos);
+
+    this.svg.margin = Utils.calculateSvgDimension(this.config.position.margin);
+    const theWidth = (this.config.position.orientation === 'vertical') ? this.svg.width : this.svg.height;
+
+    this.svg.barWidth = (theWidth - (((this.config.hours / this.config.barhours) - 1)
+                                  * this.svg.margin)) / (this.config.hours / this.config.barhours);
+    this._data = [];
+    this._bars = [];
+    this._scale = {};
+    this._needsRendering = false;
+
+    this.classes.bar = {};
+
+    this.styles.tool = {};
+    this.styles.line = {};
+    this.stylesBar = {};
+
+    if (this.dev.debug) console.log('SparkleBarChart constructor coords, dimensions', this.coords, this.dimensions, this.svg, this.config);
+  }
+
+  /** *****************************************************************************
+    * SparklineBarChartTool::computeMinMax()
+    *
+    * Summary.
+    * Compute min/max values of bars to scale them to the maximum amount.
+    *
+    */
+  computeMinMax() {
+    let min = this._series[0]; let
+      max = this._series[0];
+
+    for (let i = 1, len = this._series.length; i < len; i++) {
+      const v = this._series[i];
+      min = (v < min) ? v : min;
+      max = (v > max) ? v : max;
+    }
+    this._scale.min = min;
+    this._scale.max = max;
+    this._scale.size = (max - min);
+
+    // 2020.11.05
+    // Add 5% to the size of the scale and adjust the minimum value displayed.
+    // So every bar is displayed, instead of the min value having a bar length of zero!
+    this._scale.size = (max - min) * 1.05;
+    this._scale.min = max - this._scale.size;
+  }
+
+  /** *****************************************************************************
+    * SparklineBarChartTool::set series
+    *
+    * Summary.
+    * Sets the timeseries for the barchart tool. Is an array of states.
+    * If this is historical data, the caller has taken the time to create this.
+    * This tool only displays the result...
+    *
+    */
+  set data(states) {
+    this._series = Object.assign(states);
+    this.computeBars();
+    this._needsRendering = true;
+  }
+
+  set series(states) {
+    this._series = Object.assign(states);
+    this.computeBars();
+    this._needsRendering = true;
+  }
+
+  hasSeries() {
+    return this.defaultEntityIndex();
+  }
+
+  /** *****************************************************************************
+    * SparklineBarChartTool::computeBars()
+    *
+    * Summary.
+    * Compute start and end of bars for easy rendering.
+    *
+    */
+  computeBars({ _bars } = this) {
+    this.computeMinMax();
+
+    if (this.config.show.style === 'minmaxgradient') {
+      this.colorStopsMinMax = {};
+      this.colorStopsMinMax = {
+        [this._scale.min.toString()]: this.config.minmaxgradient.colors.min,
+        [this._scale.max.toString()]: this.config.minmaxgradient.colors.max,
+      };
+    }
+
+    // VERTICAL
+    if (this.config.position.orientation === 'vertical') {
+      if (this.dev.debug) console.log('bar is vertical');
+      this._series.forEach((item, index) => {
+        if (!_bars[index]) _bars[index] = {};
+        _bars[index].length = (this._scale.size === 0) ? 0 : ((item - this._scale.min) / (this._scale.size)) * this.svg.height;
+        _bars[index].x1 = this.svg.x + this.svg.barWidth / 2 + ((this.svg.barWidth + this.svg.margin) * index);
+        _bars[index].x2 = _bars[index].x1;
+        _bars[index].y1 = this.svg.y + this.svg.height;
+        _bars[index].y2 = _bars[index].y1 - this._bars[index].length;
+        _bars[index].dataLength = this._bars[index].length;
+      });
+      // HORIZONTAL
+    } else if (this.config.position.orientation === 'horizontal') {
+      if (this.dev.debug) console.log('bar is horizontal');
+      this._data.forEach((item, index) => {
+        if (!_bars[index]) _bars[index] = {};
+        // if (!item || isNaN(item)) item = this._scale.min;
+        _bars[index].length = (this._scale.size === 0) ? 0 : ((item - this._scale.min) / (this._scale.size)) * this.svg.width;
+        _bars[index].y1 = this.svg.y + this.svg.barWidth / 2 + ((this.svg.barWidth + this.svg.margin) * index);
+        _bars[index].y2 = _bars[index].y1;
+        _bars[index].x1 = this.svg.x;
+        _bars[index].x2 = _bars[index].x1 + this._bars[index].length;
+        _bars[index].dataLength = this._bars[index].length;
+      });
+    } else if (this.dev.debug) console.log('SparklineBarChartTool - unknown barchart orientation (horizontal or vertical)');
+  }
+
+  /** *****************************************************************************
+    * SparklineBarChartTool::_renderBars()
+    *
+    * Summary.
+    * Render all the bars. Number of bars depend on hours and barhours settings.
+    *
+    */
+  _renderBars({ _bars } = this) {
+    const svgItems = [];
+
+    if (this._bars.length === 0) return;
+
+    if (this.dev.debug) console.log('_renderBars IN', this.toolId);
+
+    this._bars.forEach((item, index) => {
+      if (this.dev.debug) console.log('_renderBars - bars', item, index);
+
+      const stroke = this.getColorFromState(this._series[index]);
+
+      if (!this.stylesBar[index])
+        this.stylesBar[index] = { ...this.config.styles.bar };
+
+      // NOTE @ 2021.10.27
+      // Lijkt dat this.classes niet gevuld wordt. geen merge enzo. is dat een bug?
+      // Nu tijdelijk opgelost door this.config te gebruiken, maar hoort niet natuurlijk als je kijkt
+      // naar de andere tools...
+
+      // Safeguard...
+      if (!(this._bars[index].y2)) console.log('sparklebarchart y2 invalid', this._bars[index]);
+      svgItems.push(svg`
+          <line id="line-segment-${this.toolId}-${index}" class="${classMap(this.config.classes.line)}"
+                    style="${styleMap(this.stylesBar[index])}"
+                    x1="${this._bars[index].x1}"
+                    x2="${this._bars[index].x2}"
+                    y1="${this._bars[index].y1}"
+                    y2="${this._bars[index].y2}"
+                    data-length="${this._bars[index].dataLength}"
+                    stroke="${stroke}"
+                    stroke-width="${this.svg.barWidth}"
+                    />
+          `);
+    });
+    if (this.dev.debug) console.log('_renderBars OUT', this.toolId);
+
+    return svg`${svgItems}`;
+  }
+
+  /** *****************************************************************************
+    * SparklineBarChartTool::render()
+    *
+    * Summary.
+    * The actual render() function called by the card for each tool.
+    *
+    */
+  render() {
+    return svg`
+        <g id="barchart-${this.toolId}" class="${classMap(this.classes.tool)}"
+          @click=${(e) => this.handleTapEvent(e, this.config)}>
+          ${this._renderBars()}
+        </g>
+      `;
+  }
+}
